services/nlp/requirements.txt
torch
transformers
spacy
sentence-transformers
elasticsearch
kafka-python
sentencepiece